Move user-proxy.service to proxy folder
