+++ src/app/components/job-state-status-bar/job-state-status-bar.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit, Input} from "@angular/core";
 import {User} from "../../models/user";
 import {Job} from "../../models/job/job";
 import {UserManager} from "../../services/user-manager.service";
-import {UserProxy} from "../../services/user-proxy.service";
+import {UserProxy} from "../../services/proxy/user-proxy.service";
 import {UserJob} from "../../models/user/user-job";
 import {JobProxy} from "../../services/job-proxy.service";
 import {Router} from "@angular/router";
