+++ src/app/services/auth-manager.service.ts
@@ -1,6 +1,6 @@
 import {Injectable, EventEmitter} from "@angular/core";
 import {User} from "../models/user";
-import {UserProxy} from "./user-proxy.service";
+import {UserProxy} from "./proxy/user-proxy.service";
 import {UserManager} from "./user-manager.service";
 
 @Injectable()
