+++ src/app/services/proxy/user-proxy.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from "@angular/core";
-import {ApiCall} from "./api-call.service";
-import {UserStatus} from "../models/user/user-status";
+import {ApiCall} from "../api-call.service";
+import {UserStatus} from "../../models/user/user-status";
 import {map} from "lodash";
-import {UserImage} from "../models/user/user-image";
-import {UserJob} from "../models/user/user-job";
+import {UserImage} from "../../models/user/user-image";
+import {UserJob} from "../../models/user/user-job";
 
 @Injectable()
 export class UserProxy {
